pyproject.toml
[tool.poetry]
name = "ml-pipeline"
version = "0.1.0"
description = "ML pipeline based on MLflow"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
mlflow = "^2.10.0"
wandb = "^0.16.3"
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"